renaming `MemberFilter` to `ChatMemberUpdated`
